requirements.txt
pdfplumber
openpyxl
python-dotenv
selenium
